Moving out ContainerNodeNameMapping to own class + Javadoc
